+++ src/microservices/Microservices.CohortPackager/Execution/ExtractJobStorage/IExtractJobStore.cs
@@ -37,7 +37,7 @@ namespace Microservices.CohortPackager.Execution.ExtractJobStorage
         /// </summary>
         /// <param name="anonVerificationMessage"></param>
         /// <param name="header"></param>
-        void PersistMessageToStore(IsIdentifiableMessage anonVerificationMessage, IMessageHeader header);
+        void PersistMessageToStore(ExtractedFileVerificationMessage anonVerificationMessage, IMessageHeader header);
 
         /// <summary>
         /// Returns a list of all jobs which are ready for final checks
@@ -86,5 +86,12 @@ namespace Microservices.CohortPackager.Execution.ExtractJobStorage
         /// <param name="jobId"></param>
         /// <returns></returns>
         IEnumerable<Tuple<string, string>> GetCompletedJobVerificationFailures(Guid jobId);
+
+        /// <summary>
+        /// Returns the full list of files that were requested but could not be produced, and a reason for each
+        /// </summary>
+        /// <param name="jobId"></param>
+        /// <returns></returns>
+        IEnumerable<string> GetCompletedJobMissingFileList(Guid jobId);
     }
 }