+++ src/microservices/Microservices.CohortPackager/Execution/ExtractJobStorage/MongoDB/ObjectModel/MongoFileStatusDoc.cs
@@ -1,7 +1,8 @@
-﻿using System;
-using JetBrains.Annotations;
+﻿using JetBrains.Annotations;
+using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
-
+using Smi.Common.Messages.Extraction;
+using System;
 
 namespace Microservices.CohortPackager.Execution.ExtractJobStorage.MongoDB.ObjectModel
 {
@@ -15,9 +16,13 @@ namespace Microservices.CohortPackager.Execution.ExtractJobStorage.MongoDB.Objec
         [NotNull]
         public MongoExtractionMessageHeaderDoc Header { get; set; }
 
-        [BsonElement("anonymisedFileName")]
+        [BsonElement("dicomFilePath")]
         [NotNull]
-        public string AnonymisedFileName { get; set; }
+        public string DicomFilePath { get; set; }
+
+        [BsonElement("outputFileName")]
+        [CanBeNull]
+        public string OutputFileName { get; set; }
 
         [BsonElement("wasAnonymised")]
         public bool WasAnonymised { get; set; }
@@ -25,22 +30,35 @@ namespace Microservices.CohortPackager.Execution.ExtractJobStorage.MongoDB.Objec
         [BsonElement("isIdentifiable")]
         public bool IsIdentifiable { get; set; }
 
+        [BsonElement("extractedFileStatus")]
+        [BsonRepresentation(BsonType.String)]
+        public ExtractedFileStatus ExtractedFileStatus { get; set; }
+
+        /// <summary>
+        /// Should only be null for identifiable extractions where the file was successfully copied. Otherwise will be the failure reason from CTP or the report content from the IsIdentifiable verification
+        /// </summary>
         [BsonElement("statusMessage")]
-        [NotNull] // NOTE(rkm 2020-02-27) Will be the failure reason from an ExtractFileStatusMessage, or the report content from an IsIdentifiableMessage
+        [CanBeNull]
         public string StatusMessage { get; set; }
 
         public MongoFileStatusDoc(
             [NotNull] MongoExtractionMessageHeaderDoc header,
-            [NotNull] string anonymisedFileName,
+            [NotNull] string dicomFilePath,
+            [CanBeNull] string outputFileName,
             bool wasAnonymised,
             bool isIdentifiable,
-            [NotNull] string statusMessage)
+            ExtractedFileStatus extractedFileStatus,
+            [CanBeNull] string statusMessage)
         {
             Header = header ?? throw new ArgumentNullException(nameof(header));
-            AnonymisedFileName = anonymisedFileName;
+            DicomFilePath = dicomFilePath ?? throw new ArgumentNullException(nameof(dicomFilePath));
+            OutputFileName = outputFileName;
             WasAnonymised = wasAnonymised;
             IsIdentifiable = isIdentifiable;
-            StatusMessage = (!string.IsNullOrWhiteSpace(statusMessage)) ? statusMessage : throw new ArgumentNullException(nameof(statusMessage));
+            ExtractedFileStatus = (extractedFileStatus != ExtractedFileStatus.Unused) ? extractedFileStatus : throw new ArgumentException(nameof(extractedFileStatus));
+            StatusMessage = statusMessage;
+            if (!IsIdentifiable && string.IsNullOrWhiteSpace(statusMessage))
+                throw new ArgumentNullException(nameof(statusMessage));
         }
 
         #region Equality Methods
@@ -48,9 +66,11 @@ namespace Microservices.CohortPackager.Execution.ExtractJobStorage.MongoDB.Objec
         protected bool Equals(MongoFileStatusDoc other)
         {
             return Equals(Header, other.Header) &&
-                   AnonymisedFileName == other.AnonymisedFileName &&
+                   DicomFilePath == other.DicomFilePath &&
+                   OutputFileName == other.OutputFileName &&
                    WasAnonymised == other.WasAnonymised &&
                    IsIdentifiable == other.IsIdentifiable &&
+                   ExtractedFileStatus == other.ExtractedFileStatus &&
                    StatusMessage == other.StatusMessage;
         }
 
@@ -71,9 +91,11 @@ namespace Microservices.CohortPackager.Execution.ExtractJobStorage.MongoDB.Objec
             unchecked
             {
                 int hashCode = (Header.GetHashCode());
-                hashCode = (hashCode * 397) ^ (AnonymisedFileName != null ? AnonymisedFileName.GetHashCode() : 0);
+                hashCode = (hashCode * 397) ^ (DicomFilePath != null ? DicomFilePath.GetHashCode() : 0);
+                hashCode = (hashCode * 397) ^ (OutputFileName != null ? OutputFileName.GetHashCode() : 0);
                 hashCode = (hashCode * 397) ^ (WasAnonymised.GetHashCode());
                 hashCode = (hashCode * 397) ^ (IsIdentifiable.GetHashCode());
+                hashCode = (hashCode * 397) ^ (ExtractedFileStatus.GetHashCode());
                 hashCode = (hashCode * 397) ^ (StatusMessage.GetHashCode());
                 return hashCode;
             }