+++ src/microservices/com.smi.microservices.ctpanonymiser/src/main/java/org/smi/ctpanonymiser/messages/ExtractedFileStatusMessage.java
@@ -3,7 +3,7 @@ package org.smi.ctpanonymiser.messages;
 import org.smi.common.messageSerialization.JsonDeserializerWithOptions.FieldRequired;
 import org.smi.common.messages.ExtractMessage;
 import org.smi.common.messages.IMessage;
-import org.smi.ctpanonymiser.util.ExtractFileStatus;
+import org.smi.ctpanonymiser.util.ExtractedFileStatus;
 
 /**
  * Message indicating the path to an anonymised file
@@ -16,7 +16,7 @@ public class ExtractedFileStatusMessage extends ExtractMessage implements IMessa
 	public String OutputFilePath;
 
     @FieldRequired
-	public ExtractFileStatus Status;
+	public ExtractedFileStatus Status;
 
 	public String StatusMessage;