+++ src/microservices/com.smi.microservices.ctpanonymiser/src/main/java/org/smi/ctpanonymiser/util/ExtractedFileStatus.java
@@ -1,19 +1,16 @@
 
 package org.smi.ctpanonymiser.util;
 
-public enum ExtractFileStatus {
-
-    Unknown,
+public enum ExtractedFileStatus {
+    /**
+	 * Unused placeholder value
+	 */
+    Unused,
 
 	/**
 	 * The file has been anonymised successfully
 	 */
 	Anonymised,
-
-	/**
-	 * The file could not be anonymised but will be retried later
-	 */
-	ErrorWillRetry,
 
 	/**
 	 * The file could not be anonymised and will not be retired