+++ src/microservices/com.smi.microservices.ctpanonymiser/src/test/java/org/smi/ctpanonymiser/test/execution/CTPAnonymiserHostTest.java
@@ -17,7 +17,7 @@ import org.smi.ctpanonymiser.Program;
 import org.smi.ctpanonymiser.execution.CTPAnonymiserHost;
 import org.smi.ctpanonymiser.messages.ExtractFileMessage;
 import org.smi.ctpanonymiser.messages.ExtractedFileStatusMessage;
-import org.smi.ctpanonymiser.util.ExtractFileStatus;
+import org.smi.ctpanonymiser.util.ExtractedFileStatus;
 
 import java.io.File;
 import java.nio.file.Paths;
@@ -192,7 +192,7 @@ public class CTPAnonymiserHostTest extends TestCase {
 
             assertEquals("FilePaths do not match", exMessage.OutputPath, recvd.OutputFilePath);
             assertEquals("Project numbers do not match", exMessage.ProjectNumber, recvd.ProjectNumber);
-            assertEquals(ExtractFileStatus.Anonymised, recvd.Status);
+            assertEquals(ExtractedFileStatus.Anonymised, recvd.Status);
         } else {
             fail("Did not receive message");
         }
@@ -241,7 +241,7 @@ public class CTPAnonymiserHostTest extends TestCase {
             _logger.info("\n" + recvd.toString());
 
             assertEquals("FilePaths do not match", null, recvd.OutputFilePath);
-            assertEquals(ExtractFileStatus.ErrorWontRetry, recvd.Status);
+            assertEquals(ExtractedFileStatus.ErrorWontRetry, recvd.Status);
         } else {
             fail("Did not receive message");
         }