+++ tests/microservices/Microservices.CohortPackager.Tests/Execution/ExtractJobStorage/ExtractJobStoreTest.cs
@@ -1,10 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using Microservices.CohortPackager.Execution.ExtractJobStorage;
+﻿using Microservices.CohortPackager.Execution.ExtractJobStorage;
 using NUnit.Framework;
 using Smi.Common.Messages;
 using Smi.Common.Messages.Extraction;
 using Smi.Common.Tests;
+using System;
+using System.Collections.Generic;
 
 
 namespace Microservices.CohortPackager.Tests.Execution.ExtractJobStorage
@@ -35,7 +35,7 @@ namespace Microservices.CohortPackager.Tests.Execution.ExtractJobStorage
             protected override void PersistMessageToStoreImpl(ExtractionRequestInfoMessage message, IMessageHeader header) { }
             protected override void PersistMessageToStoreImpl(ExtractFileCollectionInfoMessage collectionInfoMessage, IMessageHeader header) => throw new NotImplementedException();
             protected override void PersistMessageToStoreImpl(ExtractedFileStatusMessage message, IMessageHeader header) { }
-            protected override void PersistMessageToStoreImpl(IsIdentifiableMessage message, IMessageHeader header) { }
+            protected override void PersistMessageToStoreImpl(ExtractedFileVerificationMessage message, IMessageHeader header) { }
             protected override List<ExtractJobInfo> GetReadyJobsImpl(Guid specificJobId = new Guid()) => throw new NotImplementedException();
             protected override void CompleteJobImpl(Guid jobId) { }
             protected override void MarkJobFailedImpl(Guid jobId, Exception e) { }
@@ -43,6 +43,7 @@ namespace Microservices.CohortPackager.Tests.Execution.ExtractJobStorage
             protected override IEnumerable<Tuple<string, Dictionary<string, int>>> GetCompletedJobRejectionsImpl(Guid jobId) => throw new NotImplementedException();
             protected override IEnumerable<Tuple<string, string>> GetCompletedJobAnonymisationFailuresImpl(Guid jobId) => throw new NotImplementedException();
             protected override IEnumerable<Tuple<string, string>> GetCompletedJobVerificationFailuresImpl(Guid jobId) => throw new NotImplementedException();
+            protected override IEnumerable<string> GetCompletedJobMissingFileListImpl(Guid jobId) => new[] { "missing" };
         }
 
         #endregion
@@ -80,13 +81,13 @@ namespace Microservices.CohortPackager.Tests.Execution.ExtractJobStorage
             var message = new ExtractedFileStatusMessage();
             var header = new MessageHeader();
 
-            message.Status = ExtractFileStatus.Unknown;
+            message.Status = ExtractedFileStatus.Unused;
             Assert.Throws<ApplicationException>(() => testExtractJobStore.PersistMessageToStore(message, header));
 
-            message.Status = ExtractFileStatus.Anonymised;
+            message.Status = ExtractedFileStatus.Anonymised;
             Assert.Throws<ApplicationException>(() => testExtractJobStore.PersistMessageToStore(message, header));
 
-            message.Status = ExtractFileStatus.ErrorWontRetry;
+            message.Status = ExtractedFileStatus.ErrorWontRetry;
             testExtractJobStore.PersistMessageToStore(message, header);
 
         }
@@ -98,18 +99,18 @@ namespace Microservices.CohortPackager.Tests.Execution.ExtractJobStorage
             var header = new MessageHeader();
 
             // Must have AnonymisedFileName
-            var message = new IsIdentifiableMessage();
+            var message = new ExtractedFileVerificationMessage();
             message.OutputFilePath = "";
             Assert.Throws<ApplicationException>(() => testExtractJobStore.PersistMessageToStore(message, header));
 
             // Report shouldn't be an empty string or null
-            message = new IsIdentifiableMessage();
+            message = new ExtractedFileVerificationMessage();
             message.OutputFilePath = "anon.dcm";
             message.Report = "";
             Assert.Throws<ApplicationException>(() => testExtractJobStore.PersistMessageToStore(message, header));
 
             // Report needs to contain content if marked as IsIdentifiable
-            message = new IsIdentifiableMessage();
+            message = new ExtractedFileVerificationMessage();
             message.OutputFilePath = "anon.dcm";
             message.IsIdentifiable = true;
             message.Report = "[]";
@@ -119,7 +120,7 @@ namespace Microservices.CohortPackager.Tests.Execution.ExtractJobStorage
             testExtractJobStore.PersistMessageToStore(message, header);
 
             // Report can be empty if not marked as IsIdentifiable
-            message = new IsIdentifiableMessage();
+            message = new ExtractedFileVerificationMessage();
             message.OutputFilePath = "anon.dcm";
             message.IsIdentifiable = false;
             message.Report = "[]";
@@ -145,6 +146,14 @@ namespace Microservices.CohortPackager.Tests.Execution.ExtractJobStorage
             Assert.Throws<ArgumentNullException>(() => store.MarkJobFailed(Guid.NewGuid(), null));
 
             store.MarkJobFailed(Guid.NewGuid(), new Exception());
+        }
+
+        [Test]
+        public void Test_GetCompletedJobMissingFileList()
+        {
+            var store = new TestExtractJobStore();
+            Assert.Throws<ArgumentNullException>(() => store.GetCompletedJobMissingFileList(default));
+            Assert.AreEqual(new[] { "missing" }, store.GetCompletedJobMissingFileList(Guid.NewGuid()));
         }
 
         #endregion