+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-- Add SecurityCodeScan tool to build chain for .Net code
+-   Add SecurityCodeScan tool to build chain for .Net code
+-   Add identifiable extraction support
+    -   [breaking] RabbitMQ extraction config has been refactored. Queues and service config files need to be updated
+    -   [breaking] Changes to MongoDB extraction schema. Existing databases need to be updated
 
 ## [1.11.1] - 2020-08-12